Added the capability for auto detection and inclusion of clientlibs (css/less/js)
