+++ app/aem/src/server/framework-preset-aem.ts
@@ -1,7 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import * as path from 'path';
 import { Configuration } from 'webpack';
-import AemClientLibPlugin from './aem-clientlib-plugin';
 
 const modGen = (baseDir, varName, id) => {
   return `const ${varName} = require('@storybook/aem').modelProxy(${JSON.stringify(id)});`;
@@ -32,15 +31,11 @@ export function webpack(config: Configuration) {
           test: /\.content\.xml$/,
           use: [
             {
-              loader: path.resolve(__dirname, './component-loader.js'),
+              loader: path.resolve(__dirname, './aem-component-loader.js'),
             }
           ]
         }
       ],
     },
-    plugins: [
-      ...config.plugins,
-      new AemClientLibPlugin(),
-    ]
   };
 }
