+++ examples/aem-kitchen-sink/core/wcm/components/accordion/accordion.stories.js
@@ -1,8 +1,5 @@
 import exampleContent from './example_content';
 import { aemMetadata } from '@storybook/aem';
-require('./clientlibs/site/css/accordion.css');
-require('./clientlibs/site/js/accordion.js');
-require('./clientlibs/site/js/polyfills.js');
 
 export default {
   title: 'Accordion',
